Use getRelativeCropRectangleHeight to set height of content-list images
Use the  getRelativeCropRectangleHeight function within content list to dynamically set the height of images based on the cropRectangle info on the image returned.  This will dynamically set the height based on the requested width and the info returned in the image.cropRectangle{width, height}

This is currently how they are setting the primary Image on the websites.