update deps to 4.o
BaseCMS Dependencies
All core @parameter1/base-cms-* dependencies will need to be upgraded to the latest versions. Additionally, because lower-level dependencies (such as @babel/* and node-sass) had significant version adjustments, references to these should be deleted from the yarn.lock before running an install.

Action Items
Upgrade all BaseCMS dependencies to v4.x by running the below in the project root

p1-basecms-dependencies upgrade --latest
Once completed, do not run a yarn install yet. We'll save that to the very end. :)

Delete references to @babel/* dependencies in the yarn.lock file. The easiest way to do this is to search in the file using regex ^"@babel\/ and then delete all entries. There may be alot of these, and that's okay :)

Now run ./scripts/yarn.sh -- this should run in the new Node14 Docker container (which may need to be downloaded). MAKE SURE YOU RUN THE INSTALL USING THE SCRIPT This ensures the install runs in the proper version of Node.